Allow styling of the calendar header parts
This change will allow the date parts to be styled individually using the same CSS suffix as the character used in the `title_format` option. For example:

```
.pickmeup .pmu-month-text-B {
  font-size: 1.5rem;
  color: #1cc1f8;
}
.pickmeup .pmu-month-text-Y {
  font-size: 0.8rem;
  color: #ccc;
}
.pickmeup .pmu-month-text-default {
  color: red;
  text-decoration: underline;
}
```
Will produce a title like:

![screenshot](https://imgur.com/a/ZgGnWkk)

The only caveat is that CSS class selectors are case-insensitive, so `.pickmeup .pmu-month-text-B` will also match the same as `.pickmeup .pmu-month-text-b`, etc. 

I hope you like it.